# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Zed CV API
  description: AI-powered job matching platform for Zambia
  version: 0.1.0
  contact:
    email: convergeozambia@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.zedcv.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NotificationChannels:
      type: object
      properties:
        whatsapp: { type: boolean, default: true }
        email: { type: boolean, default: true }
    AutoMatchPreferences:
      type: object
      properties:
        auto_match_enabled: { type: boolean, default: true }
        notification_channels: { $ref: '#/components/schemas/NotificationChannels' }
    AutoMatchPreferencesUpdate:
      type: object
      properties:
        auto_match_enabled: { type: boolean }
        notification_channels: { $ref: '#/components/schemas/NotificationChannels' }

    # Phase 2 Initiative #4 — job-search preferences. Backs
    # /api/v1/preferences (distinct from /profile/preferences, which is
    # notification prefs). Shape mirrors
    # apps/backend/app/schemas/preferences.py and
    # apps/frontend/src/lib/api.ts::JobPreferences.
...
paths:
...
  /users/me/preferences/auto-match:
    get:
      operationId: getAutoMatchPreferences
      tags: [Users]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutoMatchPreferences' }
    patch:
      operationId: updateAutoMatchPreferences
      tags: [Users]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AutoMatchPreferencesUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutoMatchPreferences' }
...